+++ eventstore/mongodb/native/src/test/java/org/occurrent/eventstore/mongodb/nativedriver/MongoEventStoreTest.java
@@ -32,6 +32,7 @@ import io.cloudevents.CloudEvent;
 import io.cloudevents.core.builder.CloudEventBuilder;
 import io.cloudevents.core.data.PojoCloudEventData;
 import io.github.artsok.RepeatedIfExceptionsTest;
+import org.awaitility.Awaitility;
 import org.bson.Document;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.condition.EnabledForJreRange;
@@ -59,6 +60,8 @@ import java.time.ZoneId;
 import java.util.List;
 import java.util.Map;
 import java.util.*;
+import java.util.concurrent.CyclicBarrier;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -365,7 +368,7 @@ class MongoEventStoreTest {
 
             // Then
             assertAll(
-                    () ->  assertThat(writeResult.getStreamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
                     () -> assertThat(writeResult.getStreamVersion()).isEqualTo(2L)
             );
         }
@@ -384,7 +387,7 @@ class MongoEventStoreTest {
 
             // Then
             assertAll(
-                    () ->  assertThat(writeResult.getStreamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
                     () -> assertThat(writeResult.getStreamVersion()).isEqualTo(3L)
             );
         }
@@ -396,11 +399,11 @@ class MongoEventStoreTest {
 
             // Then
             assertAll(
-                    () ->  assertThat(writeResult.getStreamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
                     () -> assertThat(writeResult.getStreamVersion()).isEqualTo(0L)
             );
         }
-        
+
         @Test
         void mongo_event_store_returns_the_previous_stream_version_when_no_events_are_written_to_an_existing_stream() {
             // Given
@@ -414,12 +417,12 @@ class MongoEventStoreTest {
 
             // Then
             assertAll(
-                    () ->  assertThat(writeResult.getStreamId()).isEqualTo("name"),
+                    () -> assertThat(writeResult.getStreamId()).isEqualTo("name"),
                     () -> assertThat(writeResult.getStreamVersion()).isEqualTo(2L)
             );
         }
     }
-    
+
 
     @SuppressWarnings("ConstantConditions")
     @Nested
@@ -708,6 +711,35 @@ class MongoEventStoreTest {
     class ConditionallyWriteToMongoEventStore {
 
         LocalDateTime now = LocalDateTime.now();
+
+        @Nested
+        @DisplayName("parallel writes")
+        class ParallelWritesToEventStoreReturns {
+
+            @Test
+            void parallel_writes_to_event_store_throws_WriteConditionNotFulfilledException() {
+                // Given
+                CyclicBarrier cyclicBarrier = new CyclicBarrier(2);
+                WriteCondition writeCondition = WriteCondition.streamVersionEq(0);
+                AtomicReference<Throwable> exception = new AtomicReference<>();
+
+                // When
+                new Thread(() -> {
+                    NameDefined event = new NameDefined(UUID.randomUUID().toString(), now, "John Doe");
+                    await(cyclicBarrier);
+                    exception.set(catchThrowable(() -> persist("name", writeCondition, event)));
+                }).start();
+
+                new Thread(() -> {
+                    NameDefined event = new NameDefined(UUID.randomUUID().toString(), now, "John Doe");
+                    await(cyclicBarrier);
+                    exception.set(catchThrowable(() -> persist("name", writeCondition, event)));
+                }).start();
+
+                // Then
+                Awaitility.await().untilAsserted(() -> assertThat(exception.get()).isEqualTo(new WriteConditionNotFulfilledException("name", 0, writeCondition, "ikk")));
+            }
+        }
 
         @Nested
         @DisplayName("eq")
@@ -1753,5 +1785,13 @@ class MongoEventStoreTest {
     private MongoEventStore newMongoEventStore(TimeRepresentation timeRepresentation) {
         ConnectionString connectionString = new ConnectionString(mongoDBContainer.getReplicaSetUrl());
         return new MongoEventStore(mongoClient, connectionString.getDatabase(), "events", new EventStoreConfig(timeRepresentation));
+    }
+
+    private static void await(CyclicBarrier cyclicBarrier) {
+        try {
+            cyclicBarrier.await();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
 }